Update deps and fix broken code.

+++ AndroidApp/app/src/main/java/se/stockman/ledlamp/data/RotatingRainbbowDataObject.kt
@@ -10,8 +10,8 @@ class RotatingRainbbowDataObject(
 
     override fun toByteArray(): ByteArray {
         val bytes = ByteArray(2)
-        bytes[0] = (colorInterval * 255).toByte()
-        bytes[1] = (speed * 255).toByte()
+        bytes[0] = (colorInterval * 255).toInt().toByte()
+        bytes[1] = (speed * 255).toInt().toByte()
         return bytes
     }
